Adding a non-FCT advection option.
Non-FCT advection uses essentially the same code as FCT advection
currently. So performance is not as good as it could be.

It is bit reproducible, which is important to the ocean core.

Migrating deriv_two computation to operators.

Deriv_two computation, and some geometry routines exist in each core.
This commit migrates them from the ocean core to the shared operators
directory.

They can be found in mpas_geometry_utils and
mpas_tracer_advection_helpers now.