Adding package variables
Package variables allow a new packages attribute to be put on any var,
var_array, or var_struct in Registry.xml. The packages attribute can
only ever be specified on a persistent varaible. o

Packages also must be defined within Registry.xml. Each package gets a
logical created for it, called packageNameActive. These logicals live in
the module mpas_packages. Each core is responsible to correctly setting
these package logicals to true or false.

When the package logical is set to true, all variables that are attached
to it are allocated and allowed to be part of I/O streams. When it's
false, they aren't allocated and they won't be allowed to be part of the
I/O streams.

Logical values must be set in the routine mpas_core_setup_packages,
which is defined within the mpas_core module for each core.