Add next/prev ptrs to mulithalo exchange lists
This commit adds next and prev pointers to the mpas_multihalo_exchange_list
type.  Specifically, these are pointers to the multihalo_exchange_list
(send, recv, or copy) on the next and previous blocks on the local processor.
This commit also sets up these new pointers in mpas_block_creator.
These will be used in a future commit to write a general subroutine for
setting up communication lists.