Converting AlarmID's to string instead of integers
This allows cores and the stream manager to identify alarms on a clock
based on a string name rather than a unique integer.

Adding a kind type ShortStrKIND that is a 32 character string. This is
used for alarmIDs to keep the time spent comparing alarmIDs down.

Updating all cores to use string-based alarm IDs.