Adding alarm functionality to stream manager.
Adding the ability for alarms to be attached to streams, and for streams
to be attached to alarms. Both directions are needed to:
 * Determine which streams need to be handled when an alarm is ringing
 * Determine which alarms to check when a stream is supposed to be
   handled

Because of this, the manager has an alarms pool, which contains subpools
that represent each alarm. In each subpool is a subpool representing the
stream, given the stream's name (it's just a pointer to the stream in
the manager % streams pool). Also, the alarm subpool has a config
integer with the same name as the stream that represents the direction
the stream should be handled for this alarm.

Also, a pool representing a stream now has a subpool named alarms that
contains subpools pointing to the alarms that control the stream, and
config integers with the same names as the alarms that tell the manager
what direction the stream should associate with each alarm.