Moving the template expansion routine.
The template expansion routine use to be in stream_manager, but this
commit moves it to timekeeper, since it's more general than just the
stream manager.

Also, adding a utility routine to set the width of the year portion of
expansions. Later this width variable can be used for timestamps in
general.