Add new routine mpas_core_setup_clock() to mpas_core modules for all cores.
This new routine is responsible for initializing, via a call to mpas_create_clock(),
the main simulation clock. The routine is called immediately after the call to
mpas_core_setup_packages() in the subdriver initialization routine, but before
any fields or blocks have been allocated in the domain type. Therefore, the clock
must be set up using only information available from the namelist.