Fix pointer assignment bug in pool subroutines.
In mpas_pool_remove_config() and mpas_pool_remove_dimension(), the pointer result
of a function was being assigned to a pointer variable using '=', rather than with '=>',
leading to a write to unallocated memory.

Also, these routines were using the pointer result without first checking that it
was an associated pointer.

Minor additional cleanup: change some tab characters to spaces.