Correctly ignore XML comments in runtime stream configuration files.
The logic in src/framework/xml_stream_parser.c, function xml_syntax_check() now considers
any tag that begins with '<!' and ends with '->' to be a comment; this is not completely
correct, but it avoids more complicated logic to test on the size of the tag buffer; and,
this logic should be good enough to handle most reasonable configuration files that a user
is likely to create.

In future, it might be nice to disable checks on balanced quotes and brackets within
a comment, but for now users will simply have to balance their quotes in comments.