Enable XML comments in stream config file
Previously, a comment would be seen as unbalanced angled brackets. This
commit doesn't count <! and -> when checking for unbalanced brakcets to
allow comments.

Cases this catches as an error:
 * XML comments that span partial XML tags (or vice versa)
 * Unbalanced XML comments

NOTE: This doesn't *only* check for comments, for example

<!  test ->

is not a valid XML comment, but would not count as unbalanced brackets
by the validator.