Adding var_struct support
This commit breaks all cores from building except the sw_core, so it
should be rebased after all cores are working.

This commit adds support for listing var_structs inside a stream. A
var_struct should be specified by the name attribute. The name attribute
of var_struct blocks are required to be unique in Registry.xml now, but
they are also given a name_in_code attribute to allow the model to have
a non-unique name for var_structs.

Conflicts:
	src/driver/mpas_subdriver.F
	src/framework/Makefile
	src/framework/mpas_framework.F
	src/framework/mpas_stream_manager.F
	src/framework/xml_stream_parser.c
	src/registry/gen_inc.c