Implement the ability to attach packages to streams at runtime.
Any packages attached to default streams in the Registry.xml file will be
transferred to the runtime XML stream definition file, and users may add packages
to any stream as a semicolon-separated list, e.g.,

<stream name="output"
        type="output"
        filename_template="output.$Y-$M-$D.nc"
        records_per_file="4"
        packages="foo;blar;biff"
        output_interval="6:00:00">

</stream>

Note: the names of packages should be the names as written in the Registry.xml file,
rather than the name with the "Active" suffix as found in the registry-generated source code.

A stream will be read or written if any of the packages attached to it is active, or
if it has no attached packages; otherwise, the stream will be neither read or written.

At runtime, if the name of a package attached to a stream is not found in the
package pool for the core, a warning message is printed to standard error, but the code
will continue executing.