Reindex connectivity fields when reading and writing streams.
Immediately after reading a stream, we translate global indices to local indices
for any connectivity fields that were read by the stream.

When writing a stream, we translate local to global indices for any connectivity
fields to be written by the stream, saving a copy of the locally indexed arrays;
immediately after the stream has been written, we restore the locally indexed arrays.