Adding utility message routines
Adding an ERROR_WRITE and WARN_WRITE routine in both mpas_stream_manager
and mpas_stream_list to allow non-debugging messages to be printed
during the course of a simulation.

Specifically when an error is going to be thrown that is fatal to the
simulation, the error message will still be printed.