Bug fix in MPAS_io_set_var_indices()
If a field X with a time dimension has the same dimensions (disregarding the time dimension)
as another field Y without a time dimension, and if X and Y were decomposed differently,
the logic in MPAS_io_set_var_indices() would erroneously reuse the decomposition from Y
for the field X.

This situation can happen, e.g., when we decompose a field by simply dividing the outer dimension
by the number of MPI tasks during bootstrapping, but later decompose fields of the same dimension
differently, perhaps according to the partition file, during the model run.

The specific case that lead to the discovery of this bug occurred with the 10242-cell mesh on
12 MPI tasks, where indexToCellID was assigned 854 cells during bootstrapping, and an integer
field dimensioned (nCells Time) was later given 849 cells according to the mesh decomposition file.
The order in which the fields were processed likely played a role, causing the MPAS_io_set_var_indices()
routine to not first encounter the decomposition for indexToCellID that was created based on
the mesh decomposition file.