Add optional argument 'direction' to MPAS_stream_mgr_set_property() and MPAS_stream_mgr_get_property()
In future, it may be useful to allow properties be specific to a particular direction for a stream.
For now, the only use of the optional 'direction' argument is in retrieving the MPAS_STREAM_PROPERTY_RECORD_INTV
property, which can be different for the input and output directions of a stream.