Add logical function MPAS_stream_mgr_ringing_alarms() to the MPAS_stream_mananger module.
This new routine tests whether any I/O alarms in a stream manager are ringing; if the optional
'streamID' argument is given, only alarms for that stream are tested; if the optional argument
'direction' is given, only alarms for the specified direction are tested.

If any of the tested alarms is ringing, the function returns .true.; otherwise, it returns .false..

This routine is useful if, for example, diagnostic fields need to be computed only on timesteps when
output streams will be written; in this case, MPAS_stream_mgr_ringing_alarms() may be called to
determine whether any output streams will be written on the next call to MPAS_stream_mgr_write(),
in which case the diagnostics can be computed before calling MPAS_stream_mgr_write().