Enable single-precision output option.
Users may now select to write output files with single-precision real fields by adding
the specification

    precision="single"

to the definition of a stream in the runtime control file.

The I/O layer (in module mpas_io) will now automatically detect when real-valued
fields in a file are single-precision, and it will appropriately convert those fields
to double-precision when they are read.