Adding logic to handle single frame input files with a nearest whence
When using the NEAREST whence search while reading an input stream, if a
file with a single frame is found, previously the stream manager would
default to looking at the file before the currently file. However, it's
possible that the nearest time might be in the file after the current
file.

In this case, we compare the current time with the time read from the
current file, and if the current time is later than the read time, we
look in the next file, otherwise we look in the previous file.