Allowing stream manager to read a file with no xtime variable
This happens with some input files; when the xtime variable is not
defined there are no seekable times in the file. If the file only
contains a single time record, it will be force read. However, if the
file contains multiple time records, and whence is set to
MPAS_STREAM_EXACT_TIME the stream manager will fail to read and return
an error code.

Also, MPAS_seekStream now sets maxRecords = 0 when xtime is not found.