Fixing an issue related to duplicating var_arrays
When duplicating a field, if that field is a var_array, previously the
constituentNames attribute was not copied over properly, causing an
issue when trying to write out the var_array, or manipulate it in other
framework routines like a var_array. Sometimes this would cause a
segfault, because constituentNames was not allocated on the duplicate
var_array.