Fixing an issue with uneven block decompositions
Previously when multiple blocks are used per MPI task, if there are an
un-even number of blocks relative to the number of MPI tasks an issue
occurs when trying to determine which processor owns each block.

This commit fixes the issue and allows framework to work properly for
uneven block decompositions.