Fix for timing when code is built with single-precision RKIND.
Differencing start times and stop times in single-precision can often
lead to time intervals of zero when the start and stop times are stored
in single-precision.

To work around the issue, start and stop times are always stored in
double-precision, and their difference and sum (total_time) are also
computed in double-precision.