Merge branch 'framework/namelist_generation' into develop
This merge adds two new tools for the generation of namelist and streams files.
In addition, it migrates the ezxml library from the src/registry directory to
the external/ezxml directory.

It also adds new targets to each core.

The first target is core_input_gen, which is responsible for generating all streams
and namelist files for a core. The core is in control of where to place these where
ever they want. Within the build targets, the core has access to variables $(NL_GEN)
and $(ST_GEN) to generate namelist and streams files respectively.

The second target is post_build, which will be called after building is complete
to handle any tasks the core wants. Currently this target migrates all input files
generated during the build to the top level directory, but additional functions can
be added as the core wants.