Removing an error message from mpas_io_inq_dim
Previously, mpas_io_inq_dim wrote it's own error message when a
dimension could not be found in a file. This commit removes this error
message and returns a more useful error code to let the calling code
determine that the dimension was not found and write it's own error
message.