Merge branch 'framework/readable_dims' into develop
This merge introduces support for "readable dimensions".

This allows dimensions to be read from multiple streams during initialization, or
to have a default definition which will be used if the dimension is not read from
a file. These dimensions are only read from files if a field is also read from
the file that requires that dimension (i.e. xCell requires the nCells dimension).
Using this functionality, dimensions can be separated into relevant files, and no
longer are required to be part of the input stream used to bootstrap framework.

Conflicts:
	src/tools/registry/gen_inc.c