Updating ESM and external Makefiles
This commit allows the ESM makefile to obey rules set in a core's
build_options.mk file. In addition, it sets the MPAS_NAMELIST_SUFFIX and
MPAS_EXE_NAME as the core specifies.

It also allows a debug build of the ESM to set MPAS_DEBUG correctly.

In addition, it changes the ezxml target in the external makefile to be
ezxml-lib, which allows it to build correctly.