Modifying how non-read dimensions are handled when being decomposed
If a dimension was not read in, previously the model would halt and
throw and error message when attempting to decompose it.

This commit transitions this from being a fatal error to being a
warning, and the decomposed dimension (and corresponsing *OwnedIndices
field) are only defined if the dimension is actually read in.

If it is not read in, and a field is allocated using it, there will be a
fatal error at that point.