Adding validation of decomposed dimensions
Adding validation that causes errors when parsing registry files if a
dimension has a definition and a decomposition.

Within the model, we're currently unable to differentiate if a dimension
was decomposed or not (i.e. if a decomposed dimension is read, it's
decomposed, but if it's not read and has a definition it is not
decomposed). This causes issues with the I/O infrastructure, so
decomposed dimensions cannot have definitions as well.

This commit also cleans up some unneeded validation code.