Only consider active fields in boostrapping when determining dimensions to read from a stream.
The MPAS_stream_mgr_get_next_field() iteration method now returns all fields that belong to
a stream, regardless of whether those fields are active in the stream. To properly build a list
of dimensions to be read from a stream, we need to consider only fields that are active in
the stream in MPAS_boostrap_framework_phase2() using the optional 'isActive' argument to
MPAS_stream_mgr_get_next_field().