Generalizing MPAS framework
This commit generalizes the MPAS framework. It removes all #includes
that include code generated when parsing a Registry.xml file, and
replaces calls to generated code with function pointers that are housed
in the new core type.

A core type will be used to define everything that is specific to a
core. Inside a core type, there is a linked list of domains,
representing multiple domains that are all instances of the particular
core they live within.

NOTE: No cores build / run after this commit