Updating subdriver
The subdriver is modified to change generic subroutine calls to function
pointer calls. It's also modified to contain a core_type instance, which
houses (currently) a single domain_type instance (in a linked list). In
addition, ifdefs are added to control core specific code (like
setup_core and setup_domain).