Fixing an issue related to files with no unlimited dimension
This commit fixes an issue that prevents reading files that have no
unlimited dimension. Previously, there was incorrect logic that would
attempt to read the length of an unlimited dimension, even if one didn't
exist from a file. If the length was not read (which it wouldn't be in
the case that an unlimited dimension doesn't exist) the file handle
would be marked as invalid and no additional information could be read
from the file.

This commit changes the logic to only query the length if the dimension
id was found, and if it was not the number of records is set to -1.

NOTE: The changes in this commit depend on the undocumented behavior of
PIO to return a -1 when an unlimited dimension is not found in a netcdf
file. If this changes in the future, it will likely break support for
files without unlimited dimensions. It is unlikely that this will break
support for files with unlimited dimensions however.