Maintain iteration prev and next pointers when adding or removing pool members
The iteration_prev and iteration_next pointers, and potentially the iteration_head
and iteration_tail pointers, must be adjusted as new memers are added to the pool and
as existing items are removed from the pool. This is accomplished with some
additional code in the routines pool_add_member() and pool_remove_member().