Fix intent for output arguments in interface for xml_stream_get_attributes()
The interface in mpas_init() for the C xml_stream_get_attributes() function erroneously
declared the intent of the output arguments of this function as intent(in). However
unlikely, it is possible that with the arguments declared as intent(in), the results returned
by the function would not appear in the actual arguments, since intent(in) implies that
the value upon exit from the call should be the same as the value upon entry to the call,
i.e., the arguments are not modified by the call.