Updating how errors are handled when reading namelists
Previously, if the error code was less than 0 a message was printed that
the namelist record was not found, however this can also occur with an
improperly formatted namelist record, where values are read in on the
head node but not broadcast to other processors.

This commit changes the error handling so that values are broadcast for
error codes <= 0, and a more general error message is written.