Fixing a bug related to 1D arrays of dimensions
This commit fixes a bug that prevented getting a 1D array of dimensions
out of a pool. Previously, the contentsDim attribute of the dimension
member within the pool was given a value equal to the size of the array
passed in. When getting an array of dimensions out of a pool, an error
was thrown if the contentsDim attribute did not have a value of 1. This
prevented the array from ever being returned.

This commit sets contentsDim to 1 for all 1D arrays of dimensions that
are added to pools.