Fixing a bug related to adding 1D dimensions into a pool
This commit fixes a bug that prevents 1D dimensions from being added
into a pool.

Previously, contentsDim prevented a 1D dimension array from being
accessed inside of a pool, however the routine to add a 1D dimension
array into a pool only allocated an array as large as contentsDim. The
previous commit set this to 1, instead of `size(dims)`, which causes a
segfault in some situations when copying the values from dims into the
new array.