fix spherical triangle areas for realistic radius
Fixes an issue which prevents the correct spherical triangle area
from being obtain when working in realistic Earth spherical coordinates
with radius_sphere=6371220.

The issue was that the L’Huilier’s Formula analog to Heron's formula is
not well behaved when arguments to the tangent functions are large,
e.g., (s-ab)/2.  Normalizing coordinates to the unit sphere for use of
this formula resolves the issue.

A new unit test is introduced to enforce accuracy for this issue.