Initialize/finalize the GPTL library when MPAS is compiled with USE_GPTL=true
The initialization of the GPTL library to use the nanotime underlying timing
function takes place in mpas_init() in src/driver/mpas_subdriver.F when GPTL
is activated in the MPAS build, and the library is finalized in mpas_finalize().

Also, just before the GPTL library is finalized in mpas_finalize(), the timers
are printed for each MPI task.