Fix an issue when adding a pool to a stream related to time levels
This commit fixes an issue when adding a pool to a stream. Previously
each field in the pool was added without querying a time level. This
caused issues when running in debug mode, and adding a pool that
contained multiple time levels.

This commit makes the add_pool routine only use time level 1 when adding
a field from a pool to a stream, since the get_field routine is really
only used to ensure the field exists before adding it to a stream.

The original issue was introduced in commit 08401505