Fix initialization / read of global attributes
This commit fixes three issues related to global attributes.
 1) parent_id and mesh_spec should be read in from the input file,
    to initialize their values
 2) All global attribute values should be initialized to some value
    before trying to write the mesh.
 3) Allowing synchronization of attribute values prior to writing them
    to the output file.

When global attributes are not initialized properly, they have
(potentially) different values on each processor that will be writing a
file. If they do have different values on each processor, pnetcdf and
thus pio thinks there was an error in defining the file.

When this happens, the first field that is going to be written is
skipped, causing it to be filled with zeroes.

This commit also introduces the ability to optionally synchronize the
value of global attributes prior to writing them out. This is useful for
attributes that likely have different values on each processor, such as
file_id.