Add an isActive check when duplicating array fields
This commit fixes an issue that caused segfaults when duplicating array
fields on fields that were not active. Previously, if a field was
duplicated (for example when cloning a pool) that was decativated by a
package, the allocation of the destiation array would cause a segfault.

This commit adds a check to ensure the destination array is active
before allocating the array (since the destination array is only active
if the source array is active).

Additionally, this commit properly nullifies the array pointer of
destination fields if it is not allocated.