Add missing value to registry and attribute lists
This commit updates the definition of registry to parse the
missing_value attribute from variable definitions. The missing_value is
then provided through the field types, in case a core wants to set
elements to have this value. Additionally, missing_value is added to
each field as an attribute so external tools can access what value is
missing for each field independently.

Additionally, it adds _FillValue, and sets the defaul values of both
_FillValue and missing_value to be the NetCDF defined fill value for the
correct type.