Fix issue with splitting a string inside quotes
This commit fixes an issue where a line needs to be broken within a
quoted string, but the quoted string is ended before the max length is
reached.

For example, the if the last few characters of the string are:

exist.')

None of those represent a valid location to break the string. Some logic
was added to determine if the last space was inside quotes or not, so if
it is used to break a string it can be broken correctly.