Fix bug in assigning names to var_array constituents when packages are used
The logic for determining which constituent index to set -- or whether to set
a constituent name -- could cause segfaults when packages were used because
the incorrect variable was being checked to determine if a constituent was
active. For example, the Registry generated code like the following for each
constituent:

      if (associated(newSubPool)) then
         call mpas_pool_get_dimension(newSubPool, 'index_nc', const_index)
      end if
      if (index_counter > 0) then
         r3Ptr(2) % constituentNames(const_index) = 'nc'
      end if

If any constituents are active, then index_counter > 0, even though the constituent
being considered might not be. The correct logic should look like:

      if (associated(newSubPool)) then
         call mpas_pool_get_dimension(newSubPool, 'index_nc', const_index)
      end if
      if (const_index > 0) then
         r3Ptr(2) % constituentNames(const_index) = 'nc'
      end if

This commit simply changes 'index_counter' to 'const_index'.