Fix incorrect interval division
This commit fixes incorrect interval divisions due to the backtrack
loop. Previously, if there were an even number of divisions (i.e. 2) the
interval division routine would cause an infinite loop and/or return an
incorrect number of divisions.

This can be seen in the previous commit's test cases.

This commit fixes the issues in the interval division routine and
properly computes the number of divisions.