Fix MPAS fillvalue usage when using PIO2
The MPAS code currently relies on the availability of netcdf fillvalues
in the code. With PIO1 these fillvalues are available indirectly by
"use"ing the pio module (that "use"s the fortran netcdf module). With
PIO2 these netcdf fillvalues will no longer be available.

The solution is to use the new (currently only available in pio2) PIO*
fillvalues for int/float/double and explicitly specify the fillvalue for
char (pio2 does not support any fillvalues for chars).

PIO2 does not currently provide a PIO_FILL_CHAR variable, so this commit
specifies a character fill value of achar(0) (ASCII Null and what NetCDF
presently uses) if PIO2 is being used.  In the future, once PIO2 has been
modified to provide PIO_FILL_CHAR, this can be updated.

(This fix was proposed by jayeshkrishna in the ACME project.)