Add mpas_log module and types; add to Makefile
This commit adds a basic, rough draft for the mpas_log.F module and its
associated derived types in mpas_log_types.inc.  The commit also adds
those types to mpas_derived_types.F and adds the new module to the
framework Makefile so it gets built.  However, the new module is not
referenced anywhere from MPAS framework, driver, or cores, so it is
still "fenced off".

The log module includes preliminary API/skeletons for four public routines:
* mpas_log_init
* mpas_log_instance_create
* mpas_log_write
* mpas_log_finalize

The first two are functional with minimal features.  The latter two are
just skeletons to be fleshed out in subsequent commits.

mpas_log_instance_create is more or less mature - is simply allocates an
mpas_log_type instance and assigns initial values to all of its members.
Future changes to this routine are likely only to update membership or
default values.

mpas_log_init is a working prototype that initializes the log manager on
each task by assigning appropriate values to mpas_log_type members based
on the model configuration.  It also opens the log file(s) and sends a
"Hello World"-like message.

It is conceivable that these two routines could be merged, but I am
leaving them separate for now to provide flexibility.