Don't open a log file if optional logUnit is provided to log_init
The original implementation was attaching the file unit provided to a
new file that was then opened.  Actually, the purpose of the optional logUnit
argument is if some higher level routine has already opened a file that
is now to be used by the log module.

So this commit modifies the functionality so that if the optional
logUnit is provided, then the log module stores the unit number and
filename, and it skips the steps of defining a file name, determining
a new unit number, and opening the file.  There is also error checking
to make sure the provided unit number is attached to an open file and is
attached in an appropriate mode for the log file.