Add initial version of C wrapper for mpas_log_write( )
The routine mpas_log_write_c( ) is a wrapper function for the mpas_log_write( )
routine intended to be called from C code. In order for C to "see" this routine,
it must reside outside the mpas_log module, but is contained in mpas_log.F.

This initial version supports only a simple message. The use of variables and
argument lists (e.g., '$i' and intArgs) or specification of the message type
is not yet supported from C.

Notes for declaring and using the mpas_log_write_c( ) routine are included in
the comment header for the routine.