Add "final_only" option for the "input_interval" or "output_interval" of streams
Streams previously supported an option, "initial_only", for the "input_interval"
and "output_interval" attributes of streams that caused the stream to be read or
written only at the beginning of a simulation. This commit adds the ability to
use "final_only" for these attributes, causing the stream to be only read or
written at the end of a simulation (determined as the time when the clock reaches
MPAS_STOP_TIME).