Export MPI_COMM_SELF from mpas_dmpar module
Some parts of the MPAS framework outside of the mpas_dmpar module -- in
particular, block streams -- make use of MPI_COMM_SELF. When building with
PIO 1.x, a definition of MPI_COMM_SELF is inherited indirectly from PIO
(specifically, the pionfatt_mod module). However, when building MPAS with
PIO 2.x, this indirect inheritance cannot be relied on, and the build will
fail. This commit includes changes to explicitly make public a definition
of MPI_COMM_SELF in the mpas_dmpar module, independent of which PIO version
is being used.