Enable 64-bit integer quotients in mpas_interval_division
By making use of the ESMF capability to multiply a time intervaly by
an 8-byte integer (introduced in 44e8b7e), the mpas_interval_division
routine can now return quotients that exceed 2**31-1.

This commit also updates the call to mpas_interval_division in several
routines so that the integer quotient is declared as I8KIND kind.