Changes to framework forcing to remove need for restart timestamp file.
These changes remove the need for the forcing system to write a text file with restart timestamps for a forcing groups clock. Instead the times are stored in the regular restart file. This requires a core to define a nForcingGroupsMax dimension (larger than the number of forcing groups on a domain) and a new "forcing" pool with three variables:
<var_struct name="forcing" time_levs="1" packages="pkgForcing">
  <var name="nForcingGroupCounter" type="integer" dimensions="Time"/>
  <var name="forcingGroupNames" type="text" dimensions="nForcingGroupsMax Time"/>
  <var name="forcingGroupRestartTimes" type="text" dimensions="nForcingGroupsMax Time"/>
</var_struct>
The latter two variables must be included in the core restart stream.