Well-defined "extra" neighbors for cellsOnCells etc.
This commit creates a well-defined "extra" neighbors for the fields:
`cellsOnCell`, `edgesOnCell`, `edgesOnEdge`, and `verticiesOnCell`.

In this commit all of the well-defined extra neighbors are set to 0. A future
commit could alter the value of any of these extra neighbors by altering the
parameters `UNUSED_CELL`, `UNUSED_EDGE` or `UNUSED_VERTEX` accordingly.