Add optional argument keep_head_ptr to mpas_deallocate_field routines
In cases where we need to deallocate a field that is an element of an array
of field types, we could not pass an element of this field array to
the mpas_deallocate_field routine, since the argument was a pointer.

The field argument to the mpas_deallocate_field routines is now a target
rather than a pointer, which allows elements of arrays of fields to be
deallocated. However, since we cannot deallocate an individual element of
an array, an optional argument, keep_head_ptr, has been added. When set
to true, the mpas_deallocate_field routine will not attempt to deallocate
the argument to the routine (though the contents of that argument's field
type -- array, constituentNames, and attLists -- will be deallocated).

Now, it is possible to deallocate an array of field types, e.g.:

    type (field1DReal), dimension(:), allocatable :: field_array

    ... code that allocates the field_array ...

    do i=1,size(field_array)
        call mpas_deallocate_field(field_array(i), keep_head_ptr=.true.)
    end do
    deallocate(field_array)

*** Note: Deallocating an allocated pointer to a single field works as before,
with the exception that the actual argument to the mpas_deallocate_field routine
is not nullified (since the dummy argument is now a target and not a pointer).
Therefore, it is recommended to explicitly nullify the actual argument after
a call to mpas_deallocate_field, e.g.,

    type (field1DReal), pointer :: f

    ... code that allocates f ...

    call mpas_deallocate_field(f)
    nullify(f)

This commit also adds code to the mpas_deallocate_field routines to deallocate
the constituentNames member of the field type if is allocated.