Fix runtime deallocation error for fields with XL compilers
Prior to this commit, MPAS models would fail with the error message

   "mpas_field_routines.F", line 1981: 1525-109 Error encountered while attempting to deallocate a data object.  The program will stop.

when using the XL compilers. The issue appears to be that the XL compiler does
not allow for deallocation of dummy arguments that have the 'target' attribute.

To address these runtime deallocation errors, the mpas_deallocate_field_target
routine never deallocates the head pointer for the field; accordingly,
the keep_head_ptr optional argument has been removed.

** Note ** : The changes in this commit introduce memory leaks, since
  the mpas_deallocate_field routine calls mpas_deallocate_field_target, and only
  nullifies the head pointer of the field after these calls.