Fix pool deallocation error with XL compiler
This commit addresses an error in deallocating fields in
the mpas_pool_destroy_pool routine for fields with only one time level:

   "mpas_pool_routines.F", line 259: 1525-109 Error encountered while attempting to deallocate a data object.  The program will stop.

Fields with just one time level are allocated as an array of fields with size
one, and the first (and only) element of this array is added to a pool. During
deallocation of the fields, the above error is a result of the attempt to
deallocate an element of an array (rather than the entire array).

To address this problem, the registry-generated code to build pools allocates
fields with a single time level using a scalar field pointer, rather than
as an array of fields of size one.